Update 1.3
Added date range display to text output
Added detailed information to text output for each employee
Added output for reallocating tips
Removed percentage data
Removed option to remove record in calculator tab
Switched to decimals for improved accuracy
Changed tip reallocation to spread across all employees starting with the highest hours between employees
Changed daily tips to perform better with calculations over different lengths of time
Set columns to autofill

+++ Jamba Tips/DebugForm.cs
@@ -94,7 +94,7 @@ namespace Jamba_Tips
                     {
                         Form1.EmployeeDay employeeDay = new Form1.EmployeeDay();
                         employeeDay.name = $"Test, Employee ({i})";
-                        employeeDay.hours = parent.RoundMoney(Math.Max(2.0, rnd.NextDouble() * 8.0));
+                        employeeDay.hours = parent.RoundMoney((decimal)Math.Max(2.0, rnd.NextDouble() * 8.0));
                         employeeDay.day = parent.RoundTime(DateTime.Now.AddDays(d));
                         parent.AddDay(employeeDay);
                     }
